Moved src content to old_src to make space for code base rewriting
