+++ old_src/css/common/widget.scss
