+++ old_src/css/editor-widget.scss
