+++ old_src/css/viewer-widget.scss
