+++ old_src/pp-annotator/annotation/AnnotationViewModel.ts
