+++ old_src/pp-annotator/api/Annotation.interface.ts
