+++ old_src/pp-annotator/api/Storage.interface.ts
