+++ old_src/pp-annotator/editor/EditorContent.scss
