+++ old_src/pp-annotator/editor/EditorWidget.tsx
