+++ old_src/pp-annotator/editor/components/PriorityButton.scss
