+++ old_src/pp-annotator/editor/editor-utils.ts
