+++ old_src/pp-annotator/modules/Module.interface.ts
