+++ old_src/pp-annotator/modules/Registry.ts
