+++ old_src/pp-annotator/viewer/ViewerContentItem.tsx
