+++ old_src/pp-annotator/viewer/ViewerWidget.tsx
